Move API related files to a package folder
